server/requirements.txt
djangorestframework
django-extensions
